SocketServer::block(): avoid calling block() on closed sockets *for real*
The code was trying to do that, however due to a subtle mistake in
`SocketSever::close` the condition `!socket_isValid(i->first)` is always
false. See, that `socket_invalidate` in `SocketServer::close` assigns
the local variable, but does not alter the key in `m_socketToInfo` map.

As of now it's not a problem since calling `select` on a closed socket
returns an error. However when using `epoll_wait` to receive notifications
closing the socket removes it from the wait list, thus `epoll_wait` will
block forever.